Make RouteConfiguration into 2 classes.
ConfigureRoutes is editable, but should never be stored for reading.
ConfiguredRoutes is always immutable and stored for reading.
