+++ flow-server/src/main/java/com/vaadin/flow/router/internal/AbstractRouteRegistry.java
@@ -59,7 +59,7 @@ public abstract class AbstractRouteRegistry implements RouteRegistry {
          * @param configuration
          *         mutable routeConfiguration to make changes to
          */
-        void configure(RouteConfiguration configuration);
+        void configure(ConfigureRoutes configuration);
     }
 
     /**
@@ -74,8 +74,8 @@ public abstract class AbstractRouteRegistry implements RouteRegistry {
      * This can only be updated through {@link #configure(Configuration)} for
      * concurrency reasons.
      */
-    private volatile RouteConfiguration routeConfiguration = new RouteConfiguration();
-    private volatile RouteConfiguration editing = null;
+    private volatile ConfiguredRoutes routeConfiguration = new ConfiguredRoutes();
+    private volatile ConfigureRoutes editing = null;
 
     private CopyOnWriteArrayList<RoutesChangedListener> routesChangedListeners = new CopyOnWriteArrayList<>();
 
@@ -89,7 +89,7 @@ public abstract class AbstractRouteRegistry implements RouteRegistry {
         lock();
         try {
             if (editing == null) {
-                editing = new RouteConfiguration(routeConfiguration, true);
+                editing = new ConfigureRoutes(routeConfiguration);
             }
 
             command.configure(editing);
@@ -116,15 +116,15 @@ public abstract class AbstractRouteRegistry implements RouteRegistry {
     private void unlock() {
         if (configurationLock.getHoldCount() == 1 && editing != null) {
             try {
-                RouteConfiguration oldConfiguration = routeConfiguration;
+                ConfiguredRoutes oldConfiguration = routeConfiguration;
 
-                routeConfiguration = new RouteConfiguration(editing, false);
+                routeConfiguration = new ConfiguredRoutes(editing);
 
                 if (!routesChangedListeners.isEmpty()) {
                     List<RouteBaseData<?>> oldRoutes = flattenRoutes(
                             getRegisteredRoutes(oldConfiguration));
                     List<RouteBaseData<?>> newRoutes = flattenRoutes(
-                            getRegisteredRoutes(editing));
+                            getRegisteredRoutes(routeConfiguration));
                     List<RouteBaseData<?>> added = new ArrayList<>();
                     List<RouteBaseData<?>> removed = new ArrayList<>();
 
@@ -171,7 +171,7 @@ public abstract class AbstractRouteRegistry implements RouteRegistry {
      *
      * @return current route configuration
      */
-    protected RouteConfiguration getConfiguration() {
+    protected ConfiguredRoutes getConfiguration() {
         return routeConfiguration;
     }
 
@@ -181,7 +181,7 @@ public abstract class AbstractRouteRegistry implements RouteRegistry {
     }
 
     private List<RouteData> getRegisteredRoutes(
-            RouteConfiguration configuration) {
+            ConfiguredRoutes configuration) {
         List<RouteData> registeredRoutes = new ArrayList<>();
         configuration.getTargetRoutes().forEach((target, url) -> {
             List<Class<?>> parameters = getRouteParameters(target);
@@ -230,7 +230,7 @@ public abstract class AbstractRouteRegistry implements RouteRegistry {
     }
 
     private List<Class<? extends RouterLayout>> getParentLayouts(
-            RouteConfiguration configuration, Class<? extends Component> target,
+            ConfiguredRoutes configuration, Class<? extends Component> target,
             String url) {
         if (configuration.getRouteTarget(url) != null) {
             List<Class<? extends RouterLayout>> parentLayouts = configuration
@@ -347,8 +347,8 @@ public abstract class AbstractRouteRegistry implements RouteRegistry {
      */
     private RouteTarget addRouteToConfiguration(String path,
             Class<? extends Component> navigationTarget,
-            RouteConfiguration configuration) {
-        if (!hasLock() || !configuration.isMutable()) {
+            ConfigureRoutes configuration) {
+        if (!hasLock()) {
             throw new IllegalStateException(
                     "addRouteToConfiguration requires the registry lock and a mutable configuration.");
         }
